backend/requirements.txt
Django==5.2.7
asgiref==3.10.0
sqlparse==0.5.3
fcm-django==2.3.1
firebase_admin==7.1.0
psycopg2-binary==2.9.10
django-environ
django-cors-headers
djangorestframework==3.16.1
django-storages==1.14.6
django-storages[s3]
django-phonenumber-field[phonenumberslite]
django-resized
Pillow
gunicorn
django-probes
